Create setup.sh, one click setup for gems-of-india locally
Added setup.sh a bash script which automates entire setup of gems-of-india by checking for dependencies assuming completely new system and installs them if not present. After setting up the project it asks for user's api keys and runs the project with minimal effort. 
Made with intention to help beginners set up the project easily, Also makes it really quick for professionals to set it up.